chore(test): splitted tests
